Implementation and tests for SevenZipReader
This CL adds implementation and tests for the SevenZipReader. Adding a
new test target adds significant linking costs on all the builders, so
tests are located in //third_party/lzma_sdk, but run as part of //chrome/test:unit_tests

Implementation is heavily based on
//chrome/installer/util/lzma_util.cc, but this code should run behind
a feature flag, so we need both copies in the short term.

Bug: 1355567
Change-Id: Id1a66bc2ad122d8b475fad0c616df6d1ffdee532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3888221
Reviewed-by: thefrog <thefrog@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047629}
NOKEYCHECK=True
GitOrigin-RevId: bbe4eae196e5536f40d946f5eeae0a4179275f57