Fix PrintContext::NumberOfPages().
This one may have been behaving correctly at some point -- at least in
horizontal writing mode. In vertical writing modes, it was just broken.
Note that this function is only used by test code (both unit tests and
web tests via special APIs), though.

The assumption seems to have been that, when printing, the layout
overflow rectangle of LayoutView would be clipped in the inline
direction to not overflow the pages. This isn't the case, though, at
least not anymore. Then it compared this size to the requested page size
to deduce the scale factor that we ended up using (in case we would have
to zoom out as an attempt to prevent/reduce inline direction overflow).

The existing PrintContextTest.ScaledVerticalRL test passed by pure
accident (but not with LayoutNGPrinting enabled), because of another
problem in the code, related floating point arithmetic.
ComputePageRectsWithPageSizeInternal() adds an epsilon to the page
height, but not to the width. This was added in
https://codereview.chromium.org/1817873002 , and the assumption must
have been that it's always the height that defines the block-size of the
page. The test passes when the block-size of the element on the second
page is 10px, but most other values makes it fail (in fact, *all* the
other values that I've tried makes the test fail - only one page found,
then).

Add an epsilon to the width as well, to be consistent. This change alone
would actually make ScaledVerticalRL fail without the rest of the code
changes in this CL.

This fix makes PrintContextTest.ScaledVerticalRL pass with
LayoutNGPrinting enabled. Add more tests that would all fail without
this fix.

Bug: 1121942
Change-Id: I47a451044b739fb55719f24770c0ddde7e4b252a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3898871
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047680}
NOKEYCHECK=True
GitOrigin-RevId: daba3be8439de46c0246f6396b9aa2cbbe9e0538