Scope space shortage check to OOFs
The fix in CL:3894720 resulted in a perf regression. To fix
this, scope the space shortage check in CachedLayoutResult()
to just OOFs. This solution should likely be scoped to all
types, not just OOFs, so a TODO was added to investigate
further.

Bug: 1363898, 1362550
Change-Id: Ied2794407065fb452622e2e417f6fde152b21d84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3900063
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1047714}
NOKEYCHECK=True
GitOrigin-RevId: 9722f69fce0bf02a236d70d6f9de0d3fa92beaf8