[zlib][aosp] move chromium_zlib_hash field to end
Prior to this commit, the chromium_zlib_hash field was added to the
middle of the struct, which changes the offsets of all later fields.
This makes it more difficult for some ABI-compatibility tooling to
reconcile the upstream and chromium-modified flavors of zlib.

This commit moves the chromium_zlib_hash field to the end of the struct
so that there are no later fields.

The field was added in crrev.com/c/3596671

BUG=1316541,b/245235788
TEST=CQ

Change-Id: Id7e1756a62e55b765eeb0b24cde0c5143a715d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3897707
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Adenilson Cavalcanti <cavalcantii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047783}
NOKEYCHECK=True
GitOrigin-RevId: b67d3166f8b75dc6c04a059f553845f7d5c38b99