compute pressure: Save last reported state
As discussed in [1], any registering observer should get an update
at its sample rate. In the current implementation, if the second
registered observer has exactly the same quantization scheme as the
first one, it will be registered but will not get an update of the
state because it will be filtered by the ´PressureServiceImpl´
instance.

By introducing the ´last_reported_state_´ in the
´PressureObserverManager´ class, it is possible to update the latest
registered observer with the latest update, which is still valid.

[1] https://github.com/WICG/compute-pressure/issues/121

Bug: 1363494
Change-Id: Ie934c78b852c42668fdf5edd981a96f1ef4a3e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3892779
Reviewed-by: Fr <beaufort.francois@gmail.com>
Reviewed-by: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Commit-Queue: Arnaud Mandy <arnaud.mandy@intel.com>
Cr-Commit-Position: refs/heads/main@{#1047902}
NOKEYCHECK=True
GitOrigin-RevId: 02338f380142d998c9a40730fec2acdba8af1ec1