Use Vector::Append() to append CSS parser tokens and backing strings.
This wins back most of the performance lost (~80% of it, give or take)
in the blink_perf.css CustomPropertiesVarAlias test. Most of what's
left is due to the fact that we don't construct these very long lists
of tokens and/or backing strings (2000 of each) in place, like we could
before when just moving a vector, but it's probably not worth optimizing
for.

Change-Id: Ib8122f9575e4a66b5236742d4a2ce99b3f896b50
Bug: 1364032
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3898996
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047903}
NOKEYCHECK=True
GitOrigin-RevId: 95149ace0fcb04baf931838bab0f3d55a3c0b6f1