Avoid GetDeprecatedTaskRunner in Canvas2DLayerBridge.
Instead of having a bool to decide whether the task should be
scheduled using the IdleTaskRunner or not, adjust the tests to flush
the IdleTaskRunner. This matches what other blink code does for testing
that is dependent on idle tasks being queued and run.

BUG=624696

Change-Id: Iebec1149b9e735bdb6f9be3a064fca889b20fb49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3891840
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048051}
NOKEYCHECK=True
GitOrigin-RevId: ecd3544d8c2bd9eb76489c4ae6c414c454e2eca1