[Clipboard API] Return-early refactor in SystemClipboard
Refactored SystemClipboard to return early if the context is detached.
This CL is a follow up to crrev.com/c/3874033 where a missing check was added to custom formats' functions. No new functionality is introduced. A TODO to solve the low coverage issue is still pending and will be addressed in a future CL.

Bug: 106449
Change-Id: I2f1818005eee7408328ae07e78d0f3d8b3a8115a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3894582
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1048124}
NOKEYCHECK=True
GitOrigin-RevId: 4a4d39bcfd618d8cc29f43b83b9c9c339443d008