Fix putImageData content disappearing with OOPR-Canvas
When a putImageData call overwrites the entire canvas, it triggers the
overdraw optimization.  Because writing pixels is done out of band with
recorder draw ops, the recording were under the impression that the
canvas buffer was uninitialized after a putImageData call, which was
causing the canvas to be cleared when it should not be.

Bug: 1361165
Change-Id: I020f38e8b60da49941a52292f4f3a600d155e053
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3901091
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048132}
NOKEYCHECK=True
GitOrigin-RevId: ea26ea641d8abe9f5272d08eef8b112fb2b82501