[code-health] Remove vpython in //third_party/blink/tools
A previous attempt to change the file to vpython3
(https://crrev.com/c/3892099) was reverted because the vpython wheels
defined in the file do not match the ones used by the code. This
indicates that this file is in fact not used anywhere and the bots
using wpt_impot.py are actually using the main .vpython3 file.

The presence of a separate vpython file for this script seems to be a
remnant of the separate repository used by WebKit.

Bug: 1336295
Change-Id: Ia8f11ab2e3f849c0d9a08e1a9096d9be872424ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3900625
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Fabrice de Gans <fdegans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048134}
NOKEYCHECK=True
GitOrigin-RevId: ab45c99f3b935719e8e12ff90ca1d31bae7d3262