[TextBasedAD] Implement pause and play functionality
CL 3810947 handles the automatic pause and play of the video and audio
descriptions are not done speaking. This CL handles user-requested
pause and play.
There are several cases that are handled with this CL:
Case 1: Video and speech are in sync
- User presses pause: The video and speech will pause
- User presses play: The video and speech will resume

Case 2: Video is paused and the speech is speaking in overtime
- User presses play: The video will begin playing and the audio
  description will finish, iff there is no upcoming cue. A con to this
  approach is that the audio description may overlap with some
  dialogue.
- User is unable to press pause, as the video is already paused

Case 3: Video and speech are not in sync after play is pressed in
        case 2
- User presses pause: The video and speech will pause out of sync
- User presses play: The video and speech will resume out of sync

In all 3 cases, if an upcoming cue start time occurs while one is being
spoken, the one being spoken will be canceled to allow for the new cue
to begin at the correct time.

There is only one pause() for the media player function, so we had to
add a special case (firing_overtime) for the case that the pause() is
internally-requested (because the audio descriptions are running
overtime) and not user-requested.

Change-Id: I926e4a212c98e3c62763e8d8fac65bd7b2238238
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3863147
Commit-Queue: Aashna Sheth <aashnas@google.com>
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048322}
NOKEYCHECK=True
GitOrigin-RevId: 6c967b9beea6452cfd3155470465d4d8f272c738