Always pass the agent cluster with cloneable objects
In order for us to be able to filter MessagePort::postMessage's task
attribution ID to the current agent cluster, we need to pass along the agent cluster for such messages.
As we already have room for an agent cluster ID, this CL repurposes it.
It turns it from an optional signifying the object should be locked to
its agent cluster, to the sender's agent cluster, always present.

To make up for it, we're adding a boolean that signifies the locking
part.

This CL shouldn't have any functional changes on its own.

Bug: 1351641
Change-Id: Iae1672e06528a339fa6cc9825878fedfc0be0831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3825861
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048565}
NOKEYCHECK=True
GitOrigin-RevId: 454a9806dbd9c6fab0bc99d38d50a699038759a9