[SPC] Add locale to SecurePaymentConfirmationRequest
This CL adds the new locale hint member to the IDL file, but does not
add any specific code support for it (see the associated bug for why).

It also reorders the dictionary members to match the spec and remove the
inline documentation (as the spec/code is the appropriate place to
document), as well as adds a TODO to support the 'extensions' member.

Bug: 1365335
Change-Id: If639f51245fbb6060e998d6cc03c7edd04f9226c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3904581
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048607}
NOKEYCHECK=True
GitOrigin-RevId: 0be20eac0a91d9b962958fc6f83469507dabea54