Change CSSGridTemplatePropertyInterpolation back to experimental
There was a bug that caused incorrect interpolation at DSF != 1.

This CL sets the feature back to experimental, which I'll then merge
back to 106 (after which I'll revert this to re-enable for 108). The
bug is fixed in 107 so no action needed there.

Bug: 1356060

Change-Id: I5729582b449700f8cb44c7219897c8fd0752ff7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905624
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1048724}
NOKEYCHECK=True
GitOrigin-RevId: 903ee22bdb0094c2769e503558e019c25179997b