[code-health] Update third_party/webdriver/pylib
This is necessary to update code to be compatible with python 3. This
also adds an indication to keep the vpython wheel to the same revision
as the version pulled into third_party.
Finally, this comments out the linux_arm64 platform in the main
.vpython3 due to some wheels not being available. This makes the
presubmit check fail. It will be re-enabled once the necessary wheels
have been added.

Bug: 1364730, 1363689
Change-Id: I8cc5d613202fcee6bee15d68292a590dc2d44e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3900850
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Fabrice de Gans <fdegans@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048731}
NOKEYCHECK=True
GitOrigin-RevId: c33130cd1eaa239097d3c4a0ddbc839ea65bd8e2