cc: Add finch flag to control when deferred commits resume
r1044123 enabled the compositor to process commits without synchronizing
scroll offsets which were sent in the BeginMainFrame message. This
allowed the main frame which unblocks FCP to commit in the same main
frame event.

The above is good for FCP since the content is drawn sooner. But bad for
the FCP metric which now measures latency of the complete pipeline as
opposed just the time spent on the main thread (before UpdateLayers).

This adds a finch flag to revert to the old behaviour by default and
roll the change out via finch.

R=vmpstr@chromium.org

Bug: 1364393
Change-Id: I1ac113809739b5d7efe72ae24dd17b60a107c64a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3899855
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1048763}
NOKEYCHECK=True
GitOrigin-RevId: 1f3d1b1ca418e0df4584cf72b60e342e098ec8d8