Add fuzzer for seven zip extractor
This CL adds a fuzzer for the code in seven_zip_reader.cc. Since this
was not historically used for processing untrusted data, a fuzzer adds
confidence that this is safe to use. The fuzzer passes the input data
directly as a 7z archive, with the exception of a modified CRC table
that effectively disables CRC checks (the CRC of any data is
0). In local testing on my workstation, this fuzzer achieves
30k executions/sec.

Bug: 1355567
Change-Id: I07fed3759a16747fc6cd21a9fdaeceb435223ac2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3892476
Reviewed-by: Rohit Bhatia <bhatiarohit@google.com>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049184}
NOKEYCHECK=True
GitOrigin-RevId: 5a38277cf1ea71645ac9c00021cdbf1bbbc266ed