Don't confuse the PiP window and the opener
HTMLMediaElement will reuse the media player when moving to a same-
origin document for Document Picture in Picture.  Unfortunately, if
the player is moved back, then it would pretend that the PiP window
is the opener that's tied to the player, for lifetime purposes.

As a result, when the PiP window closes, the player would be
destroyed too.

What should happen is that the player should be tied to the
original opener, regardless of how many times it moves.

Bug: 1365706
Change-Id: Iecb5424b8381bcf9cb3f2c2d9901cc0bcf58ffc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906567
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049274}
NOKEYCHECK=True
GitOrigin-RevId: bcc08998ba177b77104863b1f39b334c2437589b