[anchor-position] Support nested block fragmentation
In a nested block fragmentation context, fragmentainers of the
inner context may not have the correct block offsets. This is
intentionally omitted because it is expensive, and it is not
necessary for existing usages.

However, when anchors in the nested multicol is fragmented by
the outer multicol, anchor query needs the offsets. This patch
changes to let anchor query to search anchors from the outer
block fragmentation context in such cases.

|NGLogicalAnchorQueryForFragmentation| can find fragments and
their offsets by traversing descendants. By giving the outer
block fragmentation context to it, it can determine the
correct offsets of the inner fragmentainers.

Bug: 1309178
Change-Id: I0e98ed8b345b3fda1a08d9bb5b4e25cb7903f306
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3881852
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1049351}
NOKEYCHECK=True
GitOrigin-RevId: 6c19acbb3773c6f9e708ce2070e24aa8046d915c