Do not use prohibited name (on a <div>) for media timeline display
In CL:1431334 we added added ARIA to our current and remaining time
display in the form of:
<div aria-label="Current time: HH:MM:SS"> and
<div aria-label="Remaining time: HH:MM:SS"> and

However, in ARIA it is illegal to put an accessible name on a generic
role, which is the computed element for a plain <div>. The reason
it is illegal is that screen readers handle it differently, in terms
of using the inner contents or outer label.

This is a precuror to help land CL:3902482, which repairs such names
and makes them descriptions, but will end up with redundant text
for the actual time.

The fix here is to remove these labels from the a11y tree via
aria-hidden, because the information is redundant with the
same information provided on the range.

However, the range also had redundant information, and the
a11y information has also been reorganized there.

Bug: 1326294
Change-Id: I2313987b0dd95e8843ecebcaf4eb4401f50e5481
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905121
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049442}
NOKEYCHECK=True
GitOrigin-RevId: 4775f014c1922ae72105c691dbd2ccf1f81b56f8