Add virtual-keyboard=resize-layout to viewport meta
Background: The Android On-Screen Keyboard (OSK) currently resizes the
entire WebContents when it shows/hides, i.e. it resizes the initial
containing block and layout viewport, causing content re-layout. When
enabled, the OSKResizesVisualViewport flag changes the OSK resize
behavior to affect only the visual viewport so that it doesn't affect
page content.

This CL adds 'virtual-keyboard' as a new key to the viewport <meta> tag.
Setting this key to 'resize-layout' will opt the page back into a mode
where the OSK resizes the initial containing block and layout viewport.
This change is behind the OSKResizeVisualViewport flag, meaning it
doesn't take effect until that flag is enabled and it currently affects
only the Android OSK.

The approach in this CL generalizses the plumbing for the
`navigator.virtualKeyboard.overlayContent` API, converting the bool to
an enum, VirtualKeyboardMode, usable by both APIs. The enum values:

* kUnset - Use the platform default behavior. On Android: currently this
          means resize the entire web contents container (i.e. affect
          page layout). When the OSKResizesVisualViewport flag is
          enabled, this value will resize only the visual viewport.

* kResizeLayout - Currently usable only on Android and only when the
          OSKResizesVisualViewport flag is enabled. When this is set,
          the keyboard will resize the entire web contents container
          (i.e. affect page layout). This is an opt-out to the
          OSKResizesVisualViewport flag, letting authors get the old
          (current) behavior when the flag is shipped.

* kOverlaysContent - The keyboard doesn't resize any viewports and sends
          `geometrychanged` web events to the page instead.

On the Android-UI side, the ApplicationViewportInsetSupplier is
responsible for insetting the visual viewport. Since this behavior can
now change dynamically, the keyboard inset suppliers need to be attached
or detached from it whenever the page changes the setting (or the page
changes). Management of this object is now moved into
CompositorViewHolder which listens to changes in the VirtualKeyboardMode
and updates the ApplicationViewportInsetSupplier and its own size
accordingly.

Bug: 1353728
Change-Id: I04e9299b1e54bcfad4adf5135521c329436a8bad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3872948
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049465}
NOKEYCHECK=True
GitOrigin-RevId: d6e44055b057810eb114be8cf2e337d557c4eb3a