Fix a DCHECK hit caused by spellchecker invocation en-entry
Previous patch (crrev.com/c/3835143) adds some flags to record the
reasons why spellchecker is triggered. The flags are cleared when the
spellchecker invocation is finished.

This creates a regression if:
- Spellchecker updates style and layout during invocation
- The layout update resets focus
- The focus reset triggers spellchecker

Then we set the trigger reason during an invocation, and then the reason
is cleared when the invocation finishes. Next time when spellchecker is
invoked, it hits a DCHECK as it can't find the trigger reason.

This patch fixes it by not triggering spellchecker if we are in the
middle of an invocation.

Fixed: 1366033
Change-Id: I06ddb1abf4fc09a6bf4b42af8b8c03e2df00e2b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3908468
Auto-Submit: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049498}
NOKEYCHECK=True
GitOrigin-RevId: f704fa58aa6a15872ba27383b2ac46dd53e8e02f