Propagate AlternateProtocolUsage enum from network service to browser and blink service.
In some cases, Chrome fails to use HTTP/3 even when the server
supports HTTP/3, so it'll be useful to show the reason in the DevTools.
The reason is to be expressed as AlternateProtocolUsage(enum), and
the function to calculate it has already been implemented in network service.(https://crrev.com/c/3897695)

However, currently it has only the enums to describe why an alternative job or a DNS HTTPS job is used, so we should make them richer.
Especially we should prepare the enum to tell why HTTP/2 is used
even when the server supports HTTP/3.

But the purpose of this CL is not to enrich the enum but to implement the function needed to propagate the enum. Specifically, we have to propagate the enum from network service to browser service and blink service in order to show it in the DevTools. This CL is in charge of that.

Design Docs is here:
https://docs.google.com/document/d/1SfMy7MMTbSn2C1RkPxnlnQdj9AnZ5-fQdEl0XO1s2ZQ/edit?usp=sharing

Bug: 1345536
Change-Id: I776e422b3462a3eb8e5f94467b6caae08c8233e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3890050
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Yoichiro Hibara <hibarayoichiro@google.com>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1050104}
NOKEYCHECK=True
GitOrigin-RevId: edf429ba1a9b4c7db472a815ff5c49c41f720577