compute pressure: Filter records in PressureObserver
This patch implements the filtering of records in the PressureObserver
and not anymore in the platform collector, for the reasons
explained below.
As described in [1]: All observers should be updated at their
samplerate unless the previous record is similar to the new record.
As discussed in [2], it shouldn't be implemented in the platform
collector.

[1]: https://github.com/WICG/compute-pressure/commit/fe277d00b98423e9b8344d666958cfda2d1a637d
[2]: https://github.com/WICG/compute-pressure/issues/121

Bug: 1359196
Change-Id: I7b926f391ac119a0d33ec4cf91738540faa039e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3868958
Reviewed-by: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Fr <beaufort.francois@gmail.com>
Commit-Queue: Arnaud Mandy <arnaud.mandy@intel.com>
Cr-Commit-Position: refs/heads/main@{#1050140}
NOKEYCHECK=True
GitOrigin-RevId: 8b1624bedd5e04c1d64295c2d5fb481c342155bd