Fix load-abort handling for <track> loading
When a load was aborted by a new load we did not stop TextTrackLoader's
|cue_load_timer_|, and could get an unexpected notification from the old
loader.

Add a new Detach() function on TextTrackLoader that cancels the load and
stops any active timer. Proactively detach and drop the reference to the
loader when a load completes. When a new load starts, and we already
have an active loader, then fail that old load - which now also detaches
from that loader, avoiding any unexpected notifications.

Add a test for the case where a pending load is aborted.

Fixed: 1365767
Change-Id: Ie4ea632d8f6273f404b75acdebde1d77e3e2cd3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3910030
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1050210}
NOKEYCHECK=True
GitOrigin-RevId: a98befc5ff65807b681efe843a86b17274f26329