Account for zoom factor in FileUploadControlIntrinsicInlineSize
In FileUploadControlIntrinsicInlineSize [1], we add
LayoutFileUploadControl::kAfterButtonSpacing to account for the
margin-inline-end of the button set at [2]. But, this fails to account
for zoom scaling, so at higher zooms the size we calculate is too
small. As a result, in some (non-English) locales the "No file chosen"
text gets cut off because we didn't make the control big enough to
fit it.

This CL fixes FileUploadControlIntrinsicInlineSize to take zoom into
account when adding this factor.

Tested manually since there doesn't seem to be a way to force
<input type="file"> into other langs in the context of a web_test.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/layout/layout_box.cc;l=254;drc=9b20e9e2991a876e34dcb4ab3d722dfb40fc43ee
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/resources/html.css;l=652?q=html.css&ss=chromium

Bug: 1365717
Change-Id: I077e7e3321904e4b1badf0c19a307cf2da77f6c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905226
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1050819}
NOKEYCHECK=True
GitOrigin-RevId: 8d7e7ac22ce02abad0aa8dfafb3bed2d7522e69d