Improve how blobs are passed over extension function responses.
Previously blobs were passed as their UUID encoded as strings in
the base::Value response data. The renderer code on receiving this
would then send back the UUID to the browser process to construct
the actual Blob object. The Blob backend however has been phasing out
usage of UUIDs, especially for passing blobs over IPC, as a major
challenge with this scheme is keeping the blobs alive in the backend
long enough for the renderer to be able to acquire a handle/reference
to the blob (resulting in flaky tests/failures with the service
worker version of the printerProvider API).

This CL changes this logic by instead passing blobs over IPC as the
actual Blob handle. This guarantees that the blob stays alive, and
eliminates the need for extra round trips and book keeping after the
response reaches the renderer.

Bug: 1196789, 1292030, 740744
Change-Id: Ifb7e235ef64fd57d3a8709ca3f54a47c22082b4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3863293
Reviewed-by: Tim <tjudkins@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1050833}
NOKEYCHECK=True
GitOrigin-RevId: aae1e8824f3109d26e7ea9ef478a70e695c3a568