COOP: fix tests using token() in Reporting-Endpoints
rodneyding@ worked on migrating the tests from using the legacy
`Report-To` header toward the standardized `Reporting-Endpoints` header.

The second headers allows defining multiple endpoints and uses
structured headers. This is a dictionary, where the keys must start with
a lower case alphabetical character. This is an issue, because we used
to use a random UUID as a key to avoid cross-test influence.

Existing tests were using a modified version of `token()` to get it
right, but it was missed during this patch:
https://github.com/web-platform-tests/wpt/pull/35824#event-7448407304

Bug: 1367599
Fixed: 1367599
Change-Id: I41023cc294164d06b9505b8d906b50215e9d8854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3918358
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051662}
NOKEYCHECK=True
GitOrigin-RevId: ae499b5a6c1ef53c3bf5ee5628c475ea284cefbd