URLPattern: Fix webidl constructor to match spec
This CL brings our webidl inline with the spec as defined in this
pull request:

https://github.com/WICG/urlpattern/pull/169

The previous webidl syntax we had was technically invalid because we
had a constructor with a trailing required URLPatternOptions.
Dictionaries with only optional attributes cannot be required as the
last argument.

There are no functional changes in this CL.

Bug: 1345036
Change-Id: I09a1771c39e416656b0196f148f58ad795077bd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3922828
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052094}
NOKEYCHECK=True
GitOrigin-RevId: 3cb4ddba6e40ea55a15cab493e1bb3b5a33b9f0a