DevTools: introduce target subtype
A subtype will let CDP client further discriminate between pages in
different states, such as prerendered frames and portals.
In addition to subtype, `inactive` property is reported for targets
that are not exposed to the user, such as BFCache or prerendered frames.

Bug: 1348385
Change-Id: I3f9b3a7c0a8c0a0c8b20af4952ceeec1a1f356d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3919724
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052155}
NOKEYCHECK=True
GitOrigin-RevId: 1acbf21b396782c38ad3accccda31c1c2dd82e22